Cuidado con el hash, puede ser que no este bien

+++ models/dni-pos.h
@@ -6,5 +6,5 @@ struct DniPos {
     uint32_t dni;
     int64_t pos; // SERA NECESARIO YA QUE NECESITAMOS MAS DE 2GB DE TAMAÑO EN LA PARTICION
 
-    DniPos(uint32_t d = 0, int64_t p = -1) : dni(d), pos(p) {}
+    DniPos(uint32_t d = 0, uint32_t p = -1) : dni(d), pos(p) {}
 };
